[backend] refactor: 초대된 팀 리스트 조회 로직 계층 분리 및 구조 개선 (Controller → Service → Repository → Mapper)

+++ backend/src/main/java/com/example/demo/member/MemberMapper.java
@@ -26,7 +26,7 @@ public interface MemberMapper {
 
     void requestInviteMember(MemberInviteInTeamRequestDTO member);   // 팀에 멤버를 초대
 
-    List<InvitedListResponse> requestInvitedList(int userId);  // 사용자가 초대된 팀 리스트 요청
+    List<InvitedListResponse> requestInvitedList(long userId);  // 사용자가 초대된 팀 리스트 요청
 
     List<TeamRequest> requestUserTeamList(int userId);  // 사용자가 멤버인 팀 리스트 요청
